pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-orchestrator"
version = "0.6.5"
description = "Lightweight CLI Agent Orchestrator for AI coding assistants"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "godofpdog"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core async framework
    "asyncio",

    # CLI interface
    "click>=8.1.0",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.0",

    # LLM clients
    "anthropic>=0.18.0",
    "openai>=1.12.0",

    # Data validation
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",

    # Configuration
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",

    # HTTP client
    "aiohttp>=3.9.0",
    "httpx>=0.26.0",

    # HTML parsing
    "beautifulsoup4>=4.12.0",

    # Utilities
    "python-dateutil>=2.8.2",
    "uuid",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",

    # Code quality
    "black>=24.1.0",
    "isort>=5.13.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",

    # Pre-commit
    "pre-commit>=3.6.0",

    # Type stubs
    "types-pyyaml>=6.0.12",
    "types-python-dateutil>=2.8.19",
]

[project.scripts]
orchestrator = "orchestrator.cli:main"

[project.urls]
Homepage = "https://github.com/godofpdog/simple_orchestrator"
Issues = "https://github.com/godofpdog/simple_orchestrator/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["orchestrator"]
skip_gitignore = true

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ARG"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=orchestrator",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
